Move bridge into react_interop.dart
